requirements.txt
Fabric==1.3.2
PIL==1.1.7
South==0.7.3
django-uni-form==0.9.0
django-celery==2.4.2
django-coverage==1.2
django-debug-toolbar==0.9.4
django-registration
django-sentry==1.8.6.2
ipython==0.11
kombu==2.1.0
paramiko==1.7.7.1
psycopg2==2.4.4
pyOpenSSL==0.12
pycrypto==2.3
pysqlite==2.6.0
sorl-thumbnail==11.09.1
django-pagination
django-crispy-forms